requirements.txt
pyzmq
python-can
numpy
PyYAML
pyserial
